feat: refactor project structure and implement proxy response handling
The project structure has been refactored, specifically impacting JSON parser and `config/links` files placement. From functional perspective, a new function is introduced to handle proxy responses based on the environmental variables set in `.env` file. This change makes the application more robust against different proxy responses and simplifies the configuration with environment variables.
