+++ .env.example
@@ -1,4 +1,7 @@
 # Application configurations
+## Reports
 CREATE_REPORT=1
-ENABLE_PROXY_REDIRECT_HANDLING=1
-PROXY_REDIRECT_HANDLING_URL=https://example.com/threat/level/10
+
+## Proxy handling
+PROXY_HANDLING=1
+PROXY_EXPECTED_RESPONSE=200
